dbusutil: move debian patch into dbustest
Newer version of go-dbus send `AUTH TEST` to the dbus test code
instead of the previous `AUTH TEST TEST`. This broke unit tests
in debian-sid and got distro patched in sid. The main snapd code
will need to get this change eventually too so integrate it now
and handle both the old and new godbus auth.