i/b/mount-control: allow custom filesystem types
Instead of hardcoding a list of allowed FS types, create a list of the
disallowed ones based on the previous review feedback from security,
which got captured in a comment. Continue using the old list of allowed
FS types as the list of FS types to be used in case the plug declaration
does not explicitly mention a FS types.

The reason for this change is that some customers are using their own
custom filesystem; since this interface is super-privileged anyway, and
any plug must be approved by the Snap Store reviewers, it doesn't seem
that this change should pose a security risk.