systemd: add `WantedBy=default.target` to snap mount units
When booting in OEM configuration mode, the systemd target is set to
`oem-config.target` instead of the usual `multi-user.target`; if our
mount units require solely the multi-user target, which is not reached
during OEM configuration mode, they won't be started.

So, let's add `default.target` to our mount units: this target is a
symbolic link to the system default target, which is usually
`multi-user.target` for a desktop session.

Most likely, there's no need to keep `multi-user.target` in the
`WantedBy` line, but since this is a hotfix, let's take a defensive
approach and keep it until we have the time to verify that removing it
does not cause any regression.

Fixes LP:#1983528