o/snapstate: do not reboot automatically on classic
Avoid rebooting automatically on classic systems when it is a modeenv
system and we are updating a essential snap. Instead, wait for the
user to perform the reboot. For that we return a retry error in the
FinishRestart function. We check if a reboot has happened by storing
the boot id of the system when the link task is performed, and then
comparing later with the current boot id in FinishRestart.