systemd: Use targets to allow systems to integrate with snap mounts
Some systems require to delay or wait for snap mounts. Adding specific
targets to group those mounts makes it possible to add `After=` or
`Before=` to all the snap mounts.

That can be useful for example to mount after `zfs-mount.service` on
some systems.

This moves the `After=zfs-mount.service` into a drop-in configuration
for `snap-mounts-pre.target`. Having seperate drop-in configuration files
allows future versions of snapd to add or remove them without having
to track the version of target files.