i/mount: merge clashing mount entries, when possible
In the case when two interfaces (or an interface and a layout) request
the same mount entry to be setup, where "same" means having the same
target, source, and FS type, let's resolve the clash by merging the two
entries into a single one.

As for the mount options, these are merged taking into account that the
snap will need to get the "union" of the permissions granted by the
merged entries: so, the resulting entry will be read/write if at least
one of the requested entries is read/write, and "rbind" will win over
"bind".