Improve and standarize debug section on tests
The idea of this change is to standarize the debug section of the tests
and also to make them more robust.

Debug is shown when the test has failed, so the preconditions for this
step are undefined, this is why we need to suppose it can fail at any
step.

In case the debug section fails for a test, then the generic debug is
not displayed and no usefull debug information is shown to understand
the problem.