cms/snap-confine: avoid race with mkdir+chown
Once snap-confine will be run as an ordinary user, the user might be
able to fiddle with our newly created SC_HOSTFS_DIR in the small
interval of time between mkdir() and chown(). To make sure this doesn't
happen, we create the directory with no permissions, and set the right
permissions only after having set the root:root ownership.

Thanks @alexmurray for the suggestion!