sandbox/cgroup: don't check V1 cgroup if V2 is active
The cgroup V1 hierarchy behaves similarly to the V2 one when it is
unmounted: the hierarchy will be visible on the filesystem while there
is at least one process using it, but it will not appear as mounted in
/proc/mounts. Even more importantly, when systemd creates a transient
scope in such a situation, the cgroup V2 will be correctly reported in
/proc/<self>/cgroup, whereas the V1 group will be

    1:name=systemd:/

(that is, the cgroup path is incorrect). In such cases, depending on the
order in which the entries appear in the /proc/<self>/cgroup file, snapd
could incorrectly think that the process has been placed in the `/`
cgroup path.

To fix the above, we consider the V1 hierarchy only if the V2 one has
not been mounted: when the V2 hierarchy is mounted, we know that this is
what will be used by us and by systemd.

Issue raised in https://forum.snapcraft.io/t/snap-freezes-after-update-to-ubuntu-22-04/31413