i/b/system_observe: honour root dir when checking for /boot/config-*
This is important for unit tests, so that we don't depend on the test
machine's setup (which might be a chroot without a /boot directory).
And in the unit test, let's create a /boot/config file ourselves: this
ensures that the interface code will create a mount entry for /boot.