i/b/system_packages_doc: restore access to LO, cups and gimp docs
Commit 5abb3aab6b74db5a9e9920c537d673806603349a broke the access to all
directories of the form /usr/share/*/* by removing the AppArmor rules
that would allow snap-update-ns to create their mimics. Specifically, it
affected:

    /usr/share/libreoffice/help
    /usr/share/cups/doc-root
    /usr/share/gimp/2.0/help

because while it allowed snap-update-ns to create the parent directory
(e.g., /usr/share/libreoffice/) it would not contain a rule which
permitted the creation of the leaf mimic.

Such a permission is only granted by explicitly passing the parent
directory to GenWritableProfile().

Fixes: LP: #1951210