interfaces: login-session-control: Allow access to all Seat Session interface members
Remove the restriction that specifies which members of the Seat and Session
D-Bus interfaces are allowed to be used. As this interface is intended to
provide a way for a snap to control login sessions, it should allow full
control of sessions, rather than just a limited subset of interface members.

The list of members is kept for the Manager interface, as that allows
operations which are out of scope of controlling the login session (such
as rebooting and powering off the host system).

Signed-off-by: Isaac True <isaac.true@canonical.com>