many: stub devicestate.Install{Finish,SetupStorageEncryption}()
This commit adds code in the snapd daemon so that the installer
API calls `devicestate.InstallFinish()` and
`devicestate.InstallSetupStorageEncryption()` when the
`/systems/<label>` POST endpoint is called.

The devicestate part will create a change that sets up the
respective action and wires the tasks into the device managers
task handlers. The tasks itself are stubs and just return an
error for now. A followup will be used to write them up.