client: prepare InstallSystemOptions for real use
This commit prepares the `InstallSystemOptions` to be ready
for real use inside the installer API. This requires two
things:

**1. Create client.InstallVolume{,Struct} **

The new struct is similar to `gadget.Volume{,Struct}` but tailored
towards the installer use-case. The `client.VolumeStructure` contains
an extra `Device` and `UnencryptedDevice` field that is filled in
from the installer.

**2. Fix type mismatch in `client.InstallSystemOptions.OnVolumes`**

The volume information in snapd is usually encoded as:
`map[string]*gadget.Volume`. However due to an oversight in the
client code (and because it was not used yet) a small mistake
made it into the code and it got encoded as:
`map[string][]gadget.Volume` there.