cgroup: Add Snap/Cgroup monitoring capabilities
This MR adds a Snap/Cgroup monitor that allows other parts of
the code to wait until all the instances of a Snap have ended
and it can be safely refreshed. It is the first step for a
bigger MR that will allow to automagically refresh a snap after
the system has notified the user that there is a refresh
available and that they must close it.