cmd/libsnap-confine-private: do not panic on chmod failure
On Debian and other distros the chmod operation on /sys/fs/bpf/snap/
fails with ENOTSUP. This means that the directory is left with the
original 0000 permissions.

Since the functionality is not affected and the permissions are more
restrictive than the ones we are attempting to set, let's not consider
this an error case, but just continue.

This fixes frequent spread failures (SNAPDENG-2205).