tests: imprve muinstaller to be more role/installer aware
The mk-classic-rootfs.sh script is either used as part of an installer
image which will have a "base.squashfs" or as part of a spread test.

For the installer image very little additional setup is needed. For
the spread test the system needs to be setup to be used from spread.

So a new "ROLE" var will be set accordingly so that the function
"prepare_classic_rootfs" from the script knows what to do.