configcore: add option to set `service.ssh.port`
This new config option allows setting the port that the ssh
daemon listens to.

Note that this option is only available on UC20+ systems because
to work during first boot it needs to write a default configuration
into the `/etc/ssh/sshd_config.d` directory. This directory is
only available in UC20+ versions of ssh.