tests: fix snap-refresh-hold if condition
A missing ! was preventing the test from failing if the auto-refresh
for test-snapd-tools_instance wasn't completed by the end of retry.
This commit also adds another check to ensure that the other snap isn't
auto-refreshed. Since this bug was discovered because the
auto-refresh was still ongoing after the retry, it's possible that even
with the fix more retries are necessary so I doubled the number.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>