Removed unsafe pointer and unneeded checks
The "unsafe" pointer seems pointless. Also, there is no way that
n is less than the size of an event, because, according to
inotify documentation, if the buffer is too small, zero will be
returned, and that's a case already being managed. Anyway, with
a buffer of 64KBytes there is always size for, at least, one
event (taking into account that NAME_MAX is 255 bytes in linux).