snap-confine: update cgroup code to run under an ordinary user
Since sna-confine is eventually going to be run under an ordinary user,
make sure that the snaps' cgroup directories are created as the root
user and group. Also, since we do not have direct control over the
ownership of the leaf files defining the cgroup (they are automatically
created using the effective user of the process that creates the
cgroup), remove the execute permission from the parent directory, so
that only the root user will be able to alter the snap's cgroup
configuration.