client: remove classic check for `snap recovery --show-keys`
The current client code has a check for `release.OnClassic` when
running: `snap recovery --show-keys`. This check is silly because
a) if anything it needs to be on the daemon side
b) now with classic+modes systems it's no longer valid

Hence this commit removes the check entirely. The error on a
classic system without encryption is still sensible:
```
$ snap recovery --show-keys
error: system does not use disk encryption
```