i/b/shutdown: allow systemctl daemon-reload
Enabling hardware with odd power managment schemes on Ubuntu Core
sometimes as a workaround requires injecting extra targets into the
systemd dependency tree, specifically as dependencies of the
reboot/poweroff/halt targets. This can be done as part of the
install-device hook script, but then the device is stuck as it needs
to call systemctl daemon-reload for the changes to take effect
before the device can reboot from install mode to run mode.

Signed-off-by: Alex Lewontin <alex.lewontin@canonical.com>