interfaces: allow wayland slot snaps to access shm files created by Firefox
Wayland clients share memory with the compositor by creating a file
descriptor, and passing it to the server via an SCM_RIGHTS message. Both
sides then mmap the fd. For GTK and Qt apps, this doesn't require any
additional permissions because they create the fd with memfd_create(),
which has no file system path.

Unfortunately Firefox's Wayland code uses shm_open(), which creates a
file system backed file. The compositor will only be able to use the fd
it receives if it's AppArmor profile allows access to that file. So we
need a rule granting access to the path it uses.