interfaces: add pwm-control interface
The pwm-control interface is similar in spirit to the gpio-control
interface, allowing access to any and all pulse-width-modulation
channels.

The interface is equally privileged. The only difference as compared to
GPIO control is the set of permissions granted. Those are modeled after
the kernel documentation referenced from a code comment.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>