gadget: refactor OnDiskStructure
Refactor OnDiskStructure by removing LaidOutStructure from it and
instead including the fields from the latter that are needed. These
fields are expected to reflect what is written in disk as opposed to
what comes in the gadget description, which might differ (for instance
in the size of the data partition).

This is a first step of the gadget refactoring, the second one will
include the on disk structure in LaidOutStructure.