daemon: remove first err logic in allLocalSnapInfos
allLocalSnapInfos was unnecessarily iterating through each of a snap's
revisions even if an error occurred. Instead of just skipping a snap
on the first error occurrence, it would accumulate info on all revisions
and then skip it anyway if an error occurred. Moreover, no caller of
allLocalSnapInfos is using the other snaps' infos (anymore) if an error
is returned so we can simplify by returning errors immediately.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>