devicestate: fix unit test failure on non-amd64 systems
The `testInstallFinishStep` test will fail on non-amd64 systems
right now because it also matches the generated kernel commandline.
This is however architecture depedant and generated in a way that
is not easily mockable (right now) because it runs as part of
the `init()` in `bootloader/assets/grub.go`.

So instead of using the exact match this commit will just check
that the common commandline (`snapd_recovery_mode=run`) is generated
and leave the exact match to the unit tests of the `assets` pkg.

This will fix the FTBFS we see in the PPAs right now.