DT-878: Original patch re-applied with conditionals
After bisecting the code, found that the hack was removed in a
commit with the text "usersession/userd: do not modify
XDG_DATA_DIRS when calling xdg-open". This sounds like a bug
fix, so this MR just applies the change only to the help://
URIs, which are the only ones that need it. It also adds a
comment explaining why XDG_DATA_DIRS is being modified, in
the hope that the code won't be removed again in the future.