o/snapshotstate: check snapshots are self-contained on import
Snapshots require administrator privileges to be imported and hence are assumed
to be trusted. However, let's try and be as helpful as possible and validate
that they are self-contained on import (ie. there are no files within the
snapshot that point outside of the parent directory) and report an error in that
case.

Thanks: Sam Sanoop <sams@snyk.io>
Signed-off-by: Alex Murray <alex.murray@canonical.com>