snap: support install/refresh that result in "Wait" changes in the cli better
On a classic system that supports mode the user experience when refreshing
a snap in wait state (like the kernel) is not great. The progress meter just
displays that a "Re-refresh" is in doing state. This is confusing for the
user as really a reboot is required.

So this commit change the behavior so that after a refresh that
ends up in a "Wait" snap will show a message:
```
Task set to wait until a manual system restart allows to continue
```