gadget,image: remove LayoutConstraints struct
LayoutConstraints was used only for the NonMBRStartOffset, which was
actually always the same value unless used by ubuntu-image. It is
possible to override this offset by setting explicitly offsets in
gadget.yaml, so we do not need it to be configurable.

When used from ubuntu-image it was being set to zero, but anyway u-i
was calculating explicitly all offsets and inserting them in
VolumeStructure, so this change should not be an issue for it.