state: move "Wait" state up in priority
This commit changes the priority of the `Wait` status to be above
the priority of Doing. The rational is that when changes are in
the `Wait` status they cannot continue unti something else has
happend. So it seems more appropriate to show this in e.g.
`snap changes`. Right now for a refresh of e.g. the kernel on
a classic system the output of snap changes is that the change
is in "doing" state which is correct but also a bit misleading
because it is waiting for the reboot. And it is in Doing state
because the "re-refresh" task is still in doing.