.shellcheckrc: ignore all unreachable code warnings
Shellcheck's unreachable code warnings are a bit over-zealous in that
they log a warning for each unreachable line instead of block. We also
sometimes leave unreachable code that is meant to be re-enabled later.
Finally, we can't disable this for specific spread tests (because
shellcheck file-wide disables must be the next non-empty line after a #!
directive), so we'd have to add a disable for each line. Therefore,
disabling for the whole project seems reasonable given the constraints
and that we sometimes leave code to be re-enabled after an exit 0.

This also removes an exit in tools/boot-state which was harmless but
unnecessary because the function invoked before always exits.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>