boot: avoid reboot loop if there is a bad try kernel (#12425)
In case a bad try kernel was in the disk (i.e. by just a simple
dangling symlink try-kernel.efi), we entered a boot loop. Avoid that.

* tests: add test that checks that there is no boot loop

when we have a dangling try-kernel.efi symlink.

* boot: return status when asking for kernel revisions

Return status when asking for kernel revisions even if there is no try
kernel. It will still be useful. Also, add some clarifying comments to
genericInitramfsSelectSnap.

* boot: clarify comments and traces

* boot: add test for bad try status with no try kernel

* cmd/snap-bootstrap: adapt error string in tests

Due to changes in bootstate.