Allow (valid) AARE expressions for posix_mq.
Currently, posix_mq is very conservative in what it allows for the
message queue names.  Only simple names, without any patterns permitted.

This is perfectly okay for some uses, however for 'Smart Background
Images', Zoom now uses posix-mq to talk between threads inside the same
snap, and it does so with dynamically generated queue names that include
PIDs.

Because it includes the PID, there is not really any static list of
queue names that we could permit and still have Zoom be functional for
this use case.

As such, we need posix-mq to allow us to specify AARE expressions, so
that we can use a wildcard such as '/us.zoom.aom.globalmgr*'.

At the same time, we can at least enforce this being a valid AARE
expression, so we do that.