tests: set runtime-timeout to 10min
The test was using 13min but that lead to errors in the journal
because the nested qemu we use is of type `-machine q35` which has
a `ICH9` chip that provides the `iTCO_wdt` watchdog timer. The max
heartbeat time is 613s on this hardware [1].

This commit changes it to 10min (600sec) and adds a check that
systemd does not error when setting the timeout. The 10min is
selected to ensure we don't run into issues from setting the timeout
too low as the test often runs on very busy vitual machines.

[1] https://www.kernel.org/doc/html/latest/watchdog/watchdog-parameters.html