cmd/snap-confine: protective measure against attacks using stdout redirection.
When debug output was enabled, due to the verbose nature, it was possible with a
small enough redirection buffer to control the flow on snap-confine, and by
doing so intercept certain operations performed by snap-confine.
To prevent this switch stdout to be non-blocking in the case where debug is enabled.