tests: test kernel/gadget refresh as part of muinstaller-real
To ensure the kernel/gadget refresh works correctly this commit adds
a new test that creates a new gadget and kernel snap and refreshes
those.

The test will also check that the reboot will not happen automatically.

Note that right now only the gadget refresh is enabled because the
kernel needs a newer version of snap-bootstrap that supports the
system-seed-null role.