workflow: show diff when the version is dirty
We had the issue that builds got reported as dirty. This got fixed in #12463

This was a bit harder to debug than needed, this commit helps to find the
reason more quickly by running the dirty detection after the
`dpkg-buildpackage` run and by fixing a bug in the extraction code in
the GH action that prevented showing the actual diff.