aspects: restore TODO about path validation
Seems like I delete a TODO about validating that paths are not subsets
of others which is still relevant. The same issue can surface in the
unsetting code since if a write is made to "foo.bar" and then "foo" is
deleted, then the entry mapped to "foo.bar" is lost.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>