snapstate,usersession: integrate refresh awareness with snapd-desktop-integration
This MR adds full integration of the new "refresh awareness"
with snapd-desktop-integration. When a running snap can be
refreshed and the "refresh awareness" function detects that
it has been closed, it sends a notification to the user to
tell they that the snap is being refreshed and they must wait
before trying to launch it. If snapd-desktop-integration is
available, a GTK notification will be shown with a progress
bar; if it isn't available, a classic FreeDesktop notification
will be shown.

When the snap has been fully refreshed, the "refreshing"
notification will be removed and a classic FreeDesktop
notification will be shown, telling the user that the snap has
been fully refreshed and that they can now launch it.