cmd/snap: introduce '--chain' flag for sign command
This flag causes 'snap sign' to append the account-key and account
assertions that correspond to the specified signing key to the signed
output of the command. This is often critical for creating working
system-user assertions, especially in air-gapped or offline situations.

The '--chain' flag takes one of three optional arguments, 'remote',
'direct', and 'local' which correspond to where it should look for the
assertion. The default is 'remote'.

Signed-off-by: Alex Lewontin <alex.lewontin@canonical.com>