cgroup: UUID generation error instead of panic
The randutil.RandomKernelUUID() used to generate a panic on failure.

The process of consolidating common code used by Pebble and Snapd
(e.g. randutil) highlighted the fact that an error return, as relied on by
Pebble, is a more flexibie approach and allows application code to decide the
severity, instead of the library function.

- Make cgroup code deal with the error that can now be returned from the UUID
  generator.

- Add unit tests to cover the new error paths in the code.