Use overlays, instead of tmpfs.
Alright, so some background on the logic here:

It is possible that some directories which we are attempting to use may
already contain files.

While it is possible that the files in question have no value, and can
simply be 'discarded' by creating a tmpfs mount on top of the directory,
it is also possible that the files are important.

This will become much more significant as we add support for nVidia DRI
files, which for reasons which will be explained in that commit, exist
inside the snap directory tree.

As such, what we really want to do is create an overlay which we can
write on, allowing us to add or replace files, without doing the wipe
which would be done with a blind tmpfs mount on the location.

This commit implements that overlay mount.