Support nVidia libva drivers.
At least zoom can use the nVidia libva acceleration if it is available,
and so we should do our best to support it.

This is significantly complicated because desktop-launcher from
snapcraft unconditionally sets LIBVA_DRIVERS_PATH, which means that we
have no choice uxcept to put any libva drivers in that directory.

This was the big driver for the switch to overlay mounts instead of
tmpfs mounts, because we are having to mangle the snap's filesystem, and
we don't want to mangle it any more than we have to.

Yes, this is a dirty hack...  But it's a dirty hack that works.