New mkimage-uc22 tool
This new tool is used to build a classic ubuntu image with kernel,
gadget snapd and base snaps.

This script is based on the one stored in the
tests/nested/manual/fde-on-classic.

This change also includes a new spread test done to validate the tool.

The new tool has new parameters which make the process easy to run in
different scenarios.