tests: fix lxd-mount-units test
Historically we had to generate a bind mount for /snap inside
lxd containers that makes /snap a "shared" mount. See LP:#1668659
for details.

Hoever it looks like lxd now sets the filesystem root in the
container to rshared and we don't (always) need this shared
mount workaround.

This commit updates the test to detect if the bind mount generation
is needed or not.