boot: do not include `base=` in modeenv for classic+modes installs
This is a followup for the PR#12532. Right now the system does
put the `base` into the modeenv on `classic+modes` system. But
here there is no need to keep track of the base as it is not
used for booting, the deb based classic system is used for that.