o/snapstate: mk pre-download tasks if running snap blocks autorefresh
If an auto-refresh is totally or partially prevented due to some snaps
being still open, create pre-download tasks so that, when the user is
asked to close the snap, it's ready to be auto-refreshed as quickly as
possible. If the inhibition window has passed, pre-downloads aren't created
and instead the user is notified that the refresh is proceeding despite
the open snap. The next PR will add the actual pre-download logic as
well as the related logic w.r.t waiting for snaps to close, concurrent
downloads, continuing delayed auto-refreshes, etc.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>