interfaces: add access to pipewire socket
This MR adds a new interface, 'pipewire-server', that gives
full access to the pipewire socket and lock file located at

    /run/user/XXX/pipewire-0
    /run/user/XXX/pipewire-0.lock

It allows not only to read and write to the socket, but also
to create it. This is needed for the following things:

    - containerize the Pipewire/Wireplumber/Pipewire-pulse daemons
    - containerize xdg-desktop-portal
    - allow to share the screen in ubuntu-core-desktop

This is NOT needed for normal use of pipewire, because
xdg-desktop-portal is able to export it in a secure maner
inside a container.

https://warthogs.atlassian.net/browse/UCD-102